+++ examples/playlist/schemas/v0-1.json
