+++ examples/playlist/testdata/valid-v1-0.json
@@ -1,14 +1,14 @@
 {
-    "apiVersion": "ext.playlists.grafana.com/v0-0",
+    "apiVersion": "playlists.ext.grafana.com/v1-0",
     "kind": "Playlist",
     "metadata": {
         "annotations": {
-            "grafana.com/createdBy": "user:1",
-            "grafana.com/updatedBy": "user:2",
+            "grafana.com/createdBy": "me",
+            "grafana.com/updatedBy": "you",
             "grafana.com/updatedTimestamp": "2023-07-26T15:54:16-07:00"
         },
         "creationTimestamp": "2023-07-26T22:54:16Z",
-        "name": "ba2eea3b-d42c-4893-b522-6bf3b67efdc5",
+        "name": "ba2eea3b",
         "namespace": "default",
         "resourceVersion": "123456",
         "uid": "63396a47-eed8-46c3-859a-5dbac7c2b241"
@@ -16,10 +16,6 @@
     "spec": {
         "interval": "100ms",
         "items": [
-            {
-                "type": "dashboard_by_id",
-                "value": "123"
-            },
             {
                 "type": "dashboard_by_tag",
                 "value": "panel-tests"
@@ -30,6 +26,6 @@
             }
         ],
         "name": "Dashboard with tag and uid (no internal id)",
-        "uid": "ba2eea3b-d42c-4893-b522-6bf3b67efdc5"
+        "xxx": "just here for the change detection version bypass"
     }
 }
